Extract Method: No Variables
This logic is based on the parser and unparser gems, which
constructs/deconstructs an abstract syntax tree. This logic constructs
a parallel tree that can be manipulated with fewer special cases.

There are really three different kinds of functions that interact with
the tree: structure, query and mutation. Structural methods make sense
directly on the object, as they are basic properties of that node in
the tree. Queries are more convenient on the object, but it was
difficult to separate the two types of functions so that type is
extracted into modules. Because mutation can change the structure of
the tree, they exist outside of it and yield a new tree rather than
directly manipulate the existing one.