Replace With TreeRewriter
I learned after trying to get a PR accepted to Unparser that it was
the wrong library to use. It was never intended to retain formatting,
instead I should be using TreeRewriter in Parser, the same way RuboCop
does.

This ended up substantially simplifying the code, so many files that I
had originally written to wrangle the syntax tree were
eliminated. There is probably more simplification to be done, but I'll
have to tackle that in subsequent changes.